Use fully qualified table names for postgres operation

+++ src/main/java/org/commcare/formplayer/sandbox/SqlHelper.java
@@ -80,8 +80,13 @@ public class SqlHelper {
         stream.println("}");
     }
 
-    public static void dropTable(Connection c, String storageKey) {
-        String sqlStatement = "DROP TABLE IF EXISTS " + storageKey;
+    public static void dropTable(Connection c, String storageKey, boolean isPostgres, String currentSchema) {
+        String sqlStatement;
+        if (isPostgres) {
+            sqlStatement = PostgresSqlHelper.getDropTableQuery(storageKey, currentSchema);
+        } else {
+            sqlStatement = "DROP TABLE IF EXISTS " + storageKey;
+        }
         try (PreparedStatement preparedStatement = c.prepareStatement(sqlStatement)) {
             preparedStatement.execute();
         } catch (SQLException e) {
@@ -139,10 +144,16 @@ public class SqlHelper {
      * Note: Caller is responsible for ensuring the prepared statement is closed
      *
      */
-    public static PreparedStatement prepareIdSelectStatement(Connection c, String storageKey, int id) {
+    public static PreparedStatement prepareIdSelectStatement(Connection c, String storageKey, int id, boolean isPostgres, String currentSchema) {
+        String sqlStatement;
+        if (isPostgres) {
+            sqlStatement = PostgresSqlHelper.getSelectStatement("*", storageKey, currentSchema);
+        } else {
+            sqlStatement = "SELECT * FROM " + storageKey;
+        }
         try {
             PreparedStatement preparedStatement =
-                    c.prepareStatement("SELECT * FROM " + storageKey + " WHERE "
+                    c.prepareStatement(sqlStatement + " WHERE "
                             + DatabaseHelper.ID_COL + " = ?;");
             preparedStatement.setInt(1, id);
             return preparedStatement;
@@ -159,7 +170,9 @@ public class SqlHelper {
      */
     public static PreparedStatement prepareTableSelectProjectionStatement(Connection c,
                                                                           String storageKey,
-                                                                          String[] projections) {
+                                                                          String[] projections,
+                                                                          boolean isPostgres,
+                                                                          String currentSchema) {
         StringBuilder builder = new StringBuilder();
         for (int i = 0; i < projections.length; i++) {
             builder.append(projections[i]);
@@ -167,9 +180,14 @@ public class SqlHelper {
                 builder.append(", ");
             }
         }
-        String queryString = "SELECT " + builder.toString() + " FROM " + storageKey + ";";
+        String sqlStatement;
+        if (isPostgres) {
+            sqlStatement = PostgresSqlHelper.getSelectStatement(builder.toString(), storageKey, currentSchema) + ";";
+        } else {
+            sqlStatement = "SELECT " + builder.toString() + " FROM " + storageKey + ";";
+        }
         try {
-            return c.prepareStatement(queryString);
+            return c.prepareStatement(sqlStatement);
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
@@ -185,7 +203,9 @@ public class SqlHelper {
     public static PreparedStatement prepareTableSelectProjectionStatement(Connection c,
                                                                           String storageKey,
                                                                           String recordId,
-                                                                          String[] projections) {
+                                                                          String[] projections,
+                                                                          boolean isPostgres,
+                                                                          String currentSchema) {
         StringBuilder builder = new StringBuilder();
         for (int i = 0; i < projections.length; i++) {
             builder.append(projections[i]);
@@ -193,9 +213,14 @@ public class SqlHelper {
                 builder.append(", ");
             }
         }
-        String queryString = "SELECT " + builder.toString() +
-                " FROM " + storageKey +
-                " WHERE " + DatabaseHelper.ID_COL + " = ?;";
+
+        String sqlStatement;
+        if (isPostgres) {
+            sqlStatement = PostgresSqlHelper.getSelectStatement(builder.toString(), storageKey, currentSchema);
+        } else {
+            sqlStatement = "SELECT " + builder.toString() + " FROM " + storageKey;
+        }
+        String queryString = sqlStatement + " WHERE " + DatabaseHelper.ID_COL + " = ?;";
         try {
             PreparedStatement preparedStatement = c.prepareStatement(queryString);
             preparedStatement.setString(1, recordId);
@@ -213,7 +238,9 @@ public class SqlHelper {
                                                                           String storageKey,
                                                                           String where,
                                                                           String values[],
-                                                                          String[] projections) {
+                                                                          String[] projections,
+                                                                          boolean isPostgres,
+                                                                          String currentSchema) {
         StringBuilder builder = new StringBuilder();
         for (int i = 0; i < projections.length; i++) {
             builder.append(projections[i]);
@@ -222,8 +249,13 @@ public class SqlHelper {
             }
         }
         try {
-            String queryString =
-                    "SELECT " + builder.toString() + " FROM " + storageKey + " WHERE " + where + ";";
+            String sqlStatement;
+            if (isPostgres) {
+                sqlStatement = PostgresSqlHelper.getSelectStatement(builder.toString(), storageKey, currentSchema);
+            } else {
+                sqlStatement = "SELECT " + builder.toString() + " FROM " + storageKey;
+            }
+            String queryString = sqlStatement + " WHERE " + where + ";";
             PreparedStatement preparedStatement = c.prepareStatement(queryString);
             for (int i = 0; i < values.length; i++) {
                 preparedStatement.setString(i + 1, values[i]);
@@ -241,11 +273,18 @@ public class SqlHelper {
     public static PreparedStatement prepareTableSelectStatement(Connection c,
                                                                 String storageKey,
                                                                 String[] fields,
-                                                                Object[] values) {
+                                                                Object[] values,
+                                                                boolean isPostgres,
+                                                                String currentSchema) {
         Pair<String, String[]> pair = DatabaseHelper.createWhere(fields, values, null);
         try {
-            String queryString =
-                    "SELECT * FROM " + storageKey + " WHERE " + pair.first + ";";
+            String sqlStatement;
+            if (isPostgres) {
+                sqlStatement = PostgresSqlHelper.getSelectStatement("*", storageKey, currentSchema);
+            } else {
+                sqlStatement = "SELECT * FROM " + storageKey;
+            }
+            String queryString = sqlStatement + " WHERE " + pair.first + ";";
             PreparedStatement preparedStatement = c.prepareStatement(queryString);
             for (int i = 0; i < pair.second.length; i++) {
                 preparedStatement.setString(i + 1, pair.second[i]);
@@ -263,10 +302,17 @@ public class SqlHelper {
     public static PreparedStatement prepareTableSelectStatement(Connection c,
                                                                 String storageKey,
                                                                 String where,
-                                                                String values[]) {
+                                                                String values[],
+                                                                boolean isPostgres,
+                                                                String currentSchema) {
         try {
-            String queryString =
-                    "SELECT * FROM " + storageKey + " WHERE " + where + ";";
+            String sqlStatement;
+            if (isPostgres) {
+                sqlStatement = PostgresSqlHelper.getSelectStatement("*", storageKey, currentSchema);
+            } else {
+                sqlStatement = "SELECT * FROM " + storageKey;
+            }
+            String queryString = sqlStatement + " WHERE " + where + ";";
             PreparedStatement preparedStatement = c.prepareStatement(queryString);
             for (int i = 0; i < values.length; i++) {
                 preparedStatement.setString(i + 1, values[i]);
@@ -294,26 +340,27 @@ public class SqlHelper {
     public static void basicInsert(Connection c,
                                    String storageKey,
                                    Map<String, Object> contentVals,
-                                   boolean isPostgres) {
+                                   boolean isPostgres,
+                                   String currentSchema) {
         Pair<List<Object>, String> valsAndInsertStatement =
-                buildInsertStatement(storageKey, contentVals);
+                buildInsertStatement(storageKey, contentVals, isPostgres, currentSchema);
         performInsert(c, valsAndInsertStatement, isPostgres);
     }
 
     public static void insertOrReplace(Connection c,
                                        String storageKey,
                                        Map<String, Object> contentValues,
-                                       boolean isPostgres) {
+                                       boolean isPostgres,
+                                       String currentSchema) {
         Pair<List<Object>, String> valsAndInsertStatement =
-                buildInsertOrReplaceStatement(storageKey, contentValues);
+                buildInsertOrReplaceStatement(storageKey, contentValues, isPostgres, currentSchema);
         performInsert(c, valsAndInsertStatement, isPostgres);
     }
 
-    private static Pair<List<Object>, String> buildInsertStatement(String storageKey,
-                                                                   Map<String, Object> contentVals,
+    private static Pair<List<Object>, String> buildInsertStatement(Map<String, Object> contentVals,
                                                                    String insertStatement) {
-        StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append(insertStatement).append(storageKey).append(" (");
+        StringBuilder stringBuilder = new StringBuilder(insertStatement);
+        stringBuilder.append(" (");
         List<Object> values = new ArrayList<>();
         String prefix = "";
         for (String key : contentVals.keySet()) {
@@ -334,13 +381,25 @@ public class SqlHelper {
     }
 
     private static Pair<List<Object>, String> buildInsertStatement(String storageKey,
-                                                                   Map<String, Object> contentVals) {
-        return buildInsertStatement(storageKey, contentVals, "INSERT INTO ");
+                                                                   Map<String, Object> contentVals,
+                                                                   boolean isPostgres,
+                                                                   String currentSchema) {
+        if (isPostgres) {
+            return buildInsertStatement(contentVals, "INSERT INTO " + currentSchema + "." + storageKey);
+        } else {
+            return buildInsertStatement(contentVals, "INSERT INTO " + storageKey);
+        }
     }
 
     private static Pair<List<Object>, String> buildInsertOrReplaceStatement(String storageKey,
-                                                                            Map<String, Object> contentVals) {
-        return buildInsertStatement(storageKey, contentVals, "INSERT OR REPLACE INTO ");
+                                                                            Map<String, Object> contentVals,
+                                                                            boolean isPostgres,
+                                                                            String currentSchema) {
+        if (isPostgres) {
+            return buildInsertStatement(contentVals, "INSERT OR REPLACE INTO " + currentSchema + "." + storageKey);
+        } else {
+            return buildInsertStatement(contentVals, "INSERT OR REPLACE INTO " + storageKey);
+        }
     }
 
     public static int insertToTable(Connection c, String storageKey, Persistable p, boolean isPostgres, String currentSchema) {
@@ -366,7 +425,7 @@ public class SqlHelper {
      * @param storageKey name of table
      * @param p          persistable to be updated
      */
-    public static void updateId(Connection c, String storageKey, Persistable p, boolean isPostgres) {
+    public static void updateId(Connection c, String storageKey, Persistable p, boolean isPostgres, String currentSchema) {
         HashMap<String, Object> map = DatabaseHelper.getMetaFieldsAndValues(p);
 
         String[] fieldNames = map.keySet().toArray(new String[map.keySet().size()]);
@@ -374,7 +433,13 @@ public class SqlHelper {
 
         Pair<String, String[]> where = org.commcare.modern.database.DatabaseHelper.createWhere(fieldNames, values, p);
 
-        String query = "UPDATE " + storageKey + " SET " + DatabaseHelper.DATA_COL + " = ? WHERE " + where.first + ";";
+        String sqlStatement;
+        if (isPostgres) {
+            sqlStatement = "UPDATE " + currentSchema + "." + storageKey;
+        } else {
+            sqlStatement = "UPDATE " + storageKey;
+        }
+        String query = sqlStatement + " SET " + DatabaseHelper.DATA_COL + " = ? WHERE " + where.first + ";";
 
         try (PreparedStatement preparedStatement = c.prepareStatement(query)){
             setPreparedStatementArgs(preparedStatement, p, where.second, isPostgres);
@@ -418,8 +483,16 @@ public class SqlHelper {
         }
     }
 
-    public static void deleteFromTableWhere(Connection connection, String tableName, String whereClause, String arg) {
-        String query = "DELETE FROM " + tableName + " WHERE " + whereClause + ";";
+    private static String getDeleteStatement(String tableName, boolean isPostgres, String currentSchema) {
+        if (isPostgres) {
+            return "DELETE FROM " + currentSchema + "." + tableName;
+        } else {
+            return "DELETE FROM " + tableName;
+        }
+    }
+
+    public static void deleteFromTableWhere(Connection connection, String tableName, String whereClause, String arg, boolean isPostgres, String currentSchema) {
+        String query = getDeleteStatement(tableName, isPostgres, currentSchema) + " WHERE " + whereClause + ";";
 
         try (PreparedStatement preparedStatement = connection.prepareStatement(query)){
             preparedStatement.setString(1, arg);
@@ -429,8 +502,8 @@ public class SqlHelper {
         }
     }
 
-    public static void deleteFromTableWhere(Connection connection, String tableName, String whereClause, String[] args) {
-        String query = "DELETE FROM " + tableName + " WHERE " + whereClause + ";";
+    public static void deleteFromTableWhere(Connection connection, String tableName, String whereClause, String[] args, boolean isPostgres, String currentSchema) {
+        String query = getDeleteStatement(tableName, isPostgres, currentSchema) + " WHERE " + whereClause + ";";
 
         try (PreparedStatement preparedStatement = connection.prepareStatement(query)){
             for (int i = 1; i <= args.length; i++) {
@@ -449,8 +522,8 @@ public class SqlHelper {
      * @param tableName  name of table
      * @param id         sql record to update
      */
-    public static void deleteIdFromTable(Connection connection, String tableName, int id) {
-        String query = "DELETE FROM " + tableName + " WHERE " + DatabaseHelper.ID_COL + " = ?;";
+    public static void deleteIdFromTable(Connection connection, String tableName, int id, boolean isPostgres, String currentSchema) {
+        String query = getDeleteStatement(tableName, isPostgres, currentSchema) + " WHERE " + DatabaseHelper.ID_COL + " = ?;";
 
         try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setInt(1, id);
@@ -466,13 +539,16 @@ public class SqlHelper {
      * @param connection Database Connection
      * @param tableName  name of table
      */
-    public static void deleteAllFromTable(Connection connection, String tableName) {
-
-        if (!isTableExist(connection, tableName)) {
+    public static void deleteAllFromTable(Connection connection, String tableName, boolean isPostgres, String currentSchema) {
+        if (isPostgres) {
+            if (!PostgresSqlHelper.isTableExists(connection, currentSchema, tableName)) {
+                return;
+            }
+        } else if (!isTableExist(connection, tableName)) {
             return;
         }
 
-        String query = "DELETE FROM " + tableName;
+        String query = getDeleteStatement(tableName, isPostgres, currentSchema) + ";";
         try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.execute();
         } catch (SQLException e) {
